+++ application/config/ion_auth.php
@@ -121,5 +121,33 @@
 	 **/
 	$config['salt_length'] = 10;
 	
+	/**
+	 * Should the salt be stored in the database?
+	 * This will change your password encryption algorithm, 
+	 * default password, 'password', changes to 
+	 * fbaa5e216d163a02ae630ab1a43372635dd374c0 with default salt.
+	 **/
+	$config['store_salt'] = false;
+	
+	/**
+	 * Message Start Delimiter
+	 **/
+	$config['message_start_delimiter'] = '<p>';
+	
+	/**
+	 * Message End Delimiter
+	 **/
+	$config['message_end_delimiter'] = '</p>';
+	
+	/**
+	 * Error Start Delimiter
+	 **/
+	$config['error_start_delimiter'] = '<p>';
+	
+	/**
+	 * Error End Delimiter
+	 **/
+	$config['error_end_delimiter'] = '</p>';
+	
 /* End of file ion_auth.php */
 /* Location: ./system/application/config/ion_auth.php */