+++ application/config/syncwiki.php
@@ -1,0 +1,11 @@
+<?php  if ( ! defined('BASEPATH')) exit('No direct script access allowed');
+
+/**
+ * Version variable, as seen in footer. Don't change this
+ */
+$config['syncwiki_version'] = '0.1.10';
+
+/**
+ * Are we using CI2? It's just for the footer again, and shouldn't be changed
+ */
+$config['syncwiki_ci2'] = true;